+++ lib/screens/auth/auth.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:login_ui/config/palette.dart';
 import 'package:login_ui/screens/auth/register.dart';
-import '../background_paint.dart';
+import '../utils/background_paint.dart';
 import 'SignIn.dart';
 
 class AuthScreen extends StatefulWidget {
