+++ lib/screens/auth/forgotpass_email.dart
@@ -7,8 +7,8 @@ import 'package:login_ui/config/database_helper.dart';
 import 'package:login_ui/config/palette.dart';
 import 'package:login_ui/screens/auth/change_password.dart';
 import '../../main.dart';
-import '../background_paint.dart';
-import 'home.dart';
+import '../utils/background_paint.dart';
+import '../home.dart';
 import 'change_password.dart';
 import 'package:http/http.dart'as http;
 import 'dart:convert';
